feat: Sync common workflows across repos
This repo hosts workflow templates for workflows that are common to
most or all repos in this organization.  This adds a new workflow here
which will automatically create PRs to sync those workflows across the
other repos we own.

For example, if a change lands in validate-pr-title.yaml here, it will
then be pushed out as PRs to all the other repos we own.