fix: Fix workflow sync failures on new forks
For some reason, the previous version of this workflow would fail to
clone a newly-forked repo.  The error was "fatal: remote error: access
denied or repository not exported".  This separates forking and
cloning into two separate steps, with a small delay in between.

This also adds code to create the PR based on the upstream main/master
instead of whatever was there at the time of the initial fork.