fix: Fix PR title check duplication, drop need for amending (#7)
Having multiple triggers for the PR title check could cause the check
to show up twice.  This could lead to one version passing and another
failing.  If the check is required, both versions had to pass, which
was a mess.  This removes the pull_request_review trigger and only
uses pull_request_target.

This also removes the config for matching the title and commit message
of single-commit PRs.  That requirement led to the need to amend
commits or add additional commits to pass the check, even if the PR
itself were properly formatted.  Now we recommend maintainers install
a particular browser extension to copy the PR title and description to
the inputs of the merge form.  If the PR is properly formatted, the
extension will make sure the final commit message matches.